General refactoring

+++ src/main/java/array/MinimumMovesToEqualArray.java
@@ -32,13 +32,13 @@ public class MinimumMovesToEqualArray {
 
     public int minMoves(int[] nums) {
         int min = nums[0];
-        int sum = 0;
+        long sum = 0;
 
         for (var num : nums) {
             min = Math.min(min, num);
             sum += num;
         }
 
-        return sum - min * nums.length;
+        return (int) (sum - (long) min * nums.length);
     }
 }
